Added DI for console application

+++ OffLogs.Business/Services/Jwt/IJwtAuthService.cs
@@ -3,7 +3,7 @@ namespace OffLogs.Business.Services.Jwt
     public interface IJwtAuthService
     {
         public string GetToken();
-        public string BuildJwt(int userId);
-        public int? GetUserId(string jwtString = null);
+        public string BuildJwt(long userId);
+        public long? GetUserId(string jwtString = null);
     }
 }
